feat: migra H2 para PostgreSQL e padroniza tratamento de exceções

+++ src/main/java/dev/java10x/CadastroDeNinjas/service/MissoesService.java
@@ -1,11 +1,13 @@
 package dev.java10x.CadastroDeNinjas.service;
 
 import dev.java10x.CadastroDeNinjas.model.MissoesModel;
+import dev.java10x.CadastroDeNinjas.model.NinjaModel;
 import dev.java10x.CadastroDeNinjas.repository.MissoesRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -17,11 +19,20 @@ public class MissoesService {
         return missoesRepository.findAll();
     }
 
+    public MissoesModel getById(Long id){
+        return missoesRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Missão com ID " + id + " não encontrado"));
+    }
+
     public MissoesModel save(MissoesModel missao){
         return missoesRepository.save(missao);
     }
 
     public void delete(Long id){
+        if(!missoesRepository.existsById(id)){
+            throw new RuntimeException("Missão com ID " + id + " não encontrado");
+        }
+
         missoesRepository.deleteById(id);
     }
 }
