+++ src/main/java/dev/java10x/CadastroDeNinjas/service/NinjaService.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -21,11 +22,20 @@ public class NinjaService {
         return ninjaRepository.findAll();
     }
 
+    public NinjaModel getById(Long id){
+        return ninjaRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Ninja com ID " + id + " não encontrado"));
+    }
+
     public NinjaModel save(NinjaModel ninja){
         return ninjaRepository.save(ninja);
     }
 
     public void delete(Long id){
+        if(!ninjaRepository.existsById(id)){
+            throw new RuntimeException("Ninja com ID " + id + " não encontrado");
+        }
+
         ninjaRepository.deleteById(id);
     }
 }
